Alter CPU-Frequencies to bring Uniformity
Signed-off-by: Shoaib0597 <Shoaib0595@gmail.com>

Earlier, the Frequencies of both the Clusters weren't uniform. For ex: big Cluster had a Frequency of 400MHz which was missing in LITTLE Cluster and LITTLE Cluster had a Frequency of 998MHz which was missing in big Cluster. This difference in Frequencies caused a few problems for those Drivers which weren't made for big.LITTLE architecture. For ex: a Thermal Driver may not be able to throttle-down the Frequencies of both the Clusters properly as it will fail finding the same Frequency in both the Clusters i.e., if the Thermal Driver tries to reduce the Frequency to 400MHz, then the big CLuster will be throttled-down properly but the same will not be the case with the LITTLE Cluster as it doesn't contains the mentioned Frequency.
This patch by me, Shoaib, addresses this issue.

Alter CPU-Frequencies to bring Uniformity (Part-2)

Signed-off-by: Shoaib0597 <Shoaib0595@gmail.com>

This patch makes all of the Frequencies of both the Clusters exactly uniform.
Further, a Mild OC of 1.6GHz on big Cluster and 1.2GHz on LITTLE Cluster is also included.
All of the Frequencies are properly set in a beautiful manner i.e., 1600MHz, 1200MHz, etc., instead of the ugly-looking 1612MHz or 1651MHz, 1209MHz, etc,.

Developed and Tested by me, Shoaib. Also, known to work on every Yureka and Yureka PLUS device irrespective of their revision.

Conflicts:
	drivers/clk/qcom/clock-gcc-8936.c